requirements.txt
flask
python-dotenv
twilio
google-generativeai
serpapi
fpdf
requests
pydub
pytest